pyproject.toml
[project]
name = "ux570-transcribe"
version = "0.1.0"
description = "Local-first transcription pipeline for Sony ICD-UX570 with optional Claude enrichment"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Eric Baruch" }]

dependencies = [
    "faster-whisper>=1.0.3",
    "typer>=0.12.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "rich>=13.7",
    "httpx>=0.27",
    "platformdirs>=4.2",
]

[project.optional-dependencies]
cloud = [
    "anthropic>=0.40.0",
    "keyring>=25.0",
]
agent = [
    "anthropic>=0.40.0",
    "keyring>=25.0",
    "pyyaml>=6.0",
]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
    "mypy>=1.10",
]
all = [
    "anthropic>=0.40.0",
    "keyring>=25.0",
    "pyyaml>=6.0",
    "mcp>=1.0.0",
]

[project.scripts]
ux570 = "ux570_transcribe.cli:app"
ux570-mcp = "ux570_transcribe.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ux570_transcribe"]

[tool.hatch.build]
include = ["src/ux570_transcribe/**", "prompts/**"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
# B008: typer requires `typer.Option(...)` in defaults — that's the documented pattern.
# RUF012: dataclass-style class-level defaults are fine here.
ignore = ["E501", "B008", "RUF012"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["RUF059"]
"src/ux570_transcribe/watch.py" = ["RUF059"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
